+++ Runtime/Scripts/VRBroadcast.cs
@@ -9,7 +9,6 @@
  */
 
 using System.Collections.Generic;
-using System.Linq;
 using UnityEngine;
 using Unity.RenderStreaming;
 using Newtonsoft.Json;
@@ -23,6 +22,9 @@ namespace FusedVR.VRStreaming {
         IOfferHandler, IAddChannelHandler, IDisconnectHandler, IDeletedConnectionHandler {
 
         #region Variables
+
+        public static VRBroadcast Instance { get; private set; } //Singleton Broadcast Manager
+
         /// <summary>
         /// Check if client is listening for the same GameID
         /// </summary>
@@ -30,19 +32,35 @@ namespace FusedVR.VRStreaming {
         [Tooltip("This ID should match the input from the Web Browser. Leave blank to match against anything")]
         private string gameID = "";
 
-        /// <summary>
-        /// Streams (video, audio, data) that need to be sent to the client
-        /// </summary>
         [SerializeField]
-        private List<Component> streams = new List<Component>();
+        [Tooltip("The maximum number of connections you would like to connect to this server.")]
+        private int maxConnections = 1;
+
+        [SerializeField]
+        [Tooltip("The player prefabs to be spawned upon a connection.")]
+        public ClientStreams[] playerPrefabs;
 
         /// <summary>
         /// List of all connectionIds that are connected
         /// </summary>
-        private List<string> connectionIds = new List<string>();
+        private Dictionary<string , ClientStreams> activeConnections = new Dictionary<string, ClientStreams>();
+
+        /// <summary>
+        /// List of all connectionIds that have connected to this server
+        /// </summary>
+        private Queue<ClientStreams> priorPlayers = new Queue<ClientStreams>();
         #endregion
 
-        #region Disconnect
+        private void Awake() {
+            //Singleton
+            if (Instance != null && Instance != this) {
+                Destroy(gameObject);
+            } else {
+                Instance = this;
+            }
+        }
+
+        #region WebRTC Events
         public void OnDeletedConnection(SignalingEventData eventData) {
             Disconnect(eventData.connectionId);
         }
@@ -52,21 +70,16 @@ namespace FusedVR.VRStreaming {
         }
 
         private void Disconnect(string connectionId) {
-            if (!connectionIds.Contains(connectionId))
+            if (!activeConnections.ContainsKey(connectionId))
                 return;
-            connectionIds.Remove(connectionId);
 
-            foreach (var source in streams.OfType<IStreamSource>()) {
-                source.SetSender(connectionId, null);
-            }
-            foreach (var receiver in streams.OfType<IStreamReceiver>()) {
-                receiver.SetReceiver(connectionId, null);
-            }
-            foreach (var channel in streams.OfType<IDataChannel>()) {
-                channel.SetChannel(connectionId, null);
-            }
+            ClientStreams client = activeConnections[connectionId];
+            client.DeleteConnection(connectionId); //remove streams
+            activeConnections.Remove(connectionId); //remove dictionary
+
+            client.gameObject.SetActive(false);
+            priorPlayers.Enqueue(client);
         }
-        #endregion
 
         /// <summary>
         /// Event that is called when an Offer is made by a client
@@ -75,40 +88,59 @@ namespace FusedVR.VRStreaming {
         public void OnOffer(SignalingEventData data) {
 
             string inputGID = GetGameID(data.sdp);
-            if ( gameID.Length != 0 && inputGID != gameID) {
+            if (gameID.Length != 0 && inputGID != gameID) {
                 Debug.Log($"Offer Doesn't Match My GameID : {inputGID}");
                 return;
             }
 
-            if (connectionIds.Count >= 1) { //if there is more than 1 connection, let's skip this offer
-                Debug.LogWarning($"Already answered this connectionId : {connectionIds[0]}");
+            if (activeConnections.Count >= maxConnections) { //if there is more than desired connection, let's skip this offer
+                Debug.LogWarning($"Reached Maxed Connections : {activeConnections.Count}");
                 return;
             }
 
-            if (connectionIds.Contains(data.connectionId)) { //if connection is already connected, skip offer
+            if (activeConnections.ContainsKey(data.connectionId)) { //if connection is already connected, skip offer
                 Debug.LogWarning($"Already answered this connectionId : {data.connectionId}");
                 return;
             }
-            connectionIds.Add(data.connectionId); //confirm we will use this connection
 
-            foreach (var source in streams.OfType<IStreamSource>()) {
-                var transceiver = AddTrack(data.connectionId, source.Track);
-                source.SetSender(data.connectionId, transceiver.Sender);
+            //only accept answer if there is a viable set of connections
+            if (playerPrefabs.Length > 0) {
+                int playerID = activeConnections.Count % playerPrefabs.Length; //get remainder for mod
+                ClientStreams player = playerPrefabs[playerID];
+
+                if (priorPlayers.Count > 0) { // if player has previously connected
+                    player = priorPlayers.Dequeue() ;
+                    player.gameObject.SetActive(true);
+                } else if (player.gameObject.scene.rootCount == 0 // if player is a prefab or more connections than prefabs
+                    || activeConnections.Count >= playerPrefabs.Length) {
+                    player = Instantiate(playerPrefabs[playerID]);
+                }
+
+                player.SetFullConnection(data.connectionId, this);
+                activeConnections.Add(data.connectionId, player); //confirm we will use this connection
+
+                SendAnswer(data.connectionId); //accept offer with an answer
             }
-            foreach (var channel in streams.OfType<IDataChannel>().Where(c => c.IsLocal)) {
-                var _channel = CreateChannel(data.connectionId, channel.Label);
-                channel.SetChannel(data.connectionId, _channel);
-            }
-            SendAnswer(data.connectionId); //accept offer with an answer
         }
 
         /// <summary>
         /// Apply Data Channel
         /// </summary>
         public void OnAddChannel(SignalingEventData data) {
-            var channel = streams.OfType<IDataChannel>().
-                FirstOrDefault(r => r.Channel == null && !r.IsLocal);
-            channel?.SetChannel(data.connectionId, data.channel);
+            if (activeConnections.ContainsKey(data.connectionId)) {
+                activeConnections[data.connectionId].SetDataChannel(data);
+            }
+        }
+
+        #endregion
+
+        /// <summary>
+        /// Broadcast Data Message to all Clients
+        /// </summary>
+        public void BroadcastDataMessage(string evt, string msg) {
+            foreach (string key in activeConnections.Keys) {
+                activeConnections[key].SendDataMessage(evt, msg); //send message to client
+            }
         }
 
         private string GetGameID(string sdp) {
@@ -121,7 +153,7 @@ namespace FusedVR.VRStreaming {
                 jsonMap.TryGetValue("user", out value);
                 return value;
             } else {
-                return ""; //TODO: check if this is a valid fail case
+                return ""; //TODO: check if this is a valid fail case as in do we want to connect to anything?
             }
 
         }